UI Improvements
-First, I updated the majority of the documentation to reflect all of
the changes that have been made.
-APA support for images
-Improved citation engine to hide random punctuation when field is
blank; it's a bit wordier, but it is going to work much better
-Improved holoribbon UX
-Added Font Awesome icons for many text-based icons
-Citation tool
-Improved table output
-Other UI touches

So far it works very well. After fixing a few more citation-related
bugs, I'm going to write out the MLA paper specs. I will write an essay
in Gltn (for my comp class), and it will be amazing.