File improvements
I've realized that localStorage definitely has a limit, and there
definitely should be a better way of saving data. :\
Also, the way that I am saving data is very inefficient. So, as the days
go on, as I work toward 1.1, expect to see many commits about storage
efficiency.

-No longer does all your metadata die when switching formats. In fact,
it will return to your file when you switch formats.
--Currently, there's a bug that prevents recovery on file load, only on
format switch
-Link to blog in 'About'
-Users must confirm to overwrite existing file on upload or on rename