Offline support and stuff
It took a lot of effort and hair pulling, but I managed to obtain
dynamic offline support. Themes and panels are installed and saved
locally, updated when online.

If you are online, then files load as usual. If offline, then a script
will execute that loads the offline version and acts rather
frictionless.

Now just fine-tuning the AppCache and it should be good.

-More theme support in Gltn
-Better error handling
-Better save support (but somehow screws up randomly anyway >:(
-Better offline integration
-Better panels support for 3rd parties
-Cleaned up panel animations and UI
-Created a new Translate Panel that is available in the store using
Google Translate in order to test out the panel fixes
-Fixed logo