Context API
Style API renamed to remove any confusion
While hovertags work with it now, and it will markup stuff and easily go
to the notes, typing is awkward.

Due to continually adding/removing spans, your cursor freaks out. This
is alleviated on one side through the findTextReplaceText() function,
but that only works for text. To remove HTML and restore the cursor, I
need to use Rangy.

So far, that works in isolated experiments, but I've yet to get success
in integrating the system into Gltn. I've been narrowing it down for the
past two hours, and I think I'm getting closer, but I'm not there yet.

Hopefully inspiration strikes tomorrow, or this isn't going to launch
soon.