Foundation
Moved stuff to the Foundation Framework for responsiveness  and better
inherit design. This means that a lot of things are going to change
under the hood, and the docs will really need to be updated. (Can GitHub
support richer docs or should that be custom?)
-A lot of CSS has been updated
-Columns
-Switched to Foundation Reveal modals instead of custom implementation
-Metadata length now responsive
-Inserting objects now more reliable
-Accurate word count
-Ribbon modified, moved to unique script for Foundation UI
-Theme now has a new palette property
-There are now a ton of buttons and some button classes to target
-Panels now animate much nicer
-Updated included themes to integrate with the Foundation modifications

More under the hood stuff to come