1.3.1
Gltn has jumped a little, and this commit actually has a bunch of stuff
added.

There's plenty more issues to tackle.

I can close #7 (Dictionary Previous Searches) There's an API for it
(documented), and it shows up in the dictionary panel, letting you
easily get back to a previous word.

#21 could be closed (Dictionary Class), although documentation...

Close #33 (Theme Panel UI) with a bunch of new Theme stuff I've been
building
Close #34 (settings break on JSON strings) Now that it works, it allows
for much richer data to be saved
Close #35 (Dictionary Panel UX) with a bunch of ui adjustments
Close #36 (Save Theme Settings) with a new class structure for Themes
I can also close #37 and close #38 in the same vein. Hot swapping themes
is already here! It's not perfect, but it's certainly cool.