Progress
Finally started to narrow down on that \frac bug which has plagued me
every since the feature was available.

I have a few small things left to fix with 1.3.2. Otherwise, I'm ready
for the next milestone which is albeit, much more significant and
complicated.

I can close #50, close #69, and close #30