Spreadsheets
Spreadsheets are now integrated very well into Gltn. They form, users
can enter text, numbers, simple formulas, or Spreadsheet functions. This
can close #128, close #134, and close #137

Next is #130 and #123 for Spreadsheets. #126 is also important
(documenting all of this too).

There's also collaboration as key things for this release. The rest can
be on the backburner.